fix coordinates

+++ modules/gui/screen.h
@@ -12,14 +12,14 @@ struct Screen {
 };
 
 struct Point2D {
-    unsigned int x;
-    unsigned int y;
+    float x;
+    float y;
 
-    Point2D operator-(const Point2D &right) const{
+    Point2D operator-(const Point2D &right) const {
         return {this->x - right.x, this->y - right.y};
     }
 
-    Point2D operator/(int right) const{
+    Point2D operator/(int right) const {
         return {this->x / right, this->y / right};
     }
 };
